passage des dégats et de la vie en double pour pouvoir etre plus fin et upgrade statistique de base opérationel via les touche 1 2 3 4 5 (temporaire en attandant le menu)

+++ P2Java/src/meteorsiege/gameitems/GameItemInterface.java
@@ -7,8 +7,8 @@ public interface GameItemInterface
 	{
 	public void draw(Graphics g);
 	public void nextPosition(int deltaTime);
-	public int getDamage();
-	public boolean takeDamage(int value);
+	public double getDamage();
+	public boolean takeDamage(double value);
 	public void destroy();
 	public int getReward();
 	}
